# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 实验室智能助手 API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    业务 API 的唯一机器可读契约。除登录和健康检查外，所有接口要求 Bearer Token。
    时间字段必须是带时区偏移的 ISO-8601 字符串；服务端持久化时转换为 UTC。
servers:
  - url: /api
security:
  - bearerAuth: []
...
paths:
...
  /labs:
    get:
      tags: [Labs]
      operationId: listLabs
      summary: 分页查询实验室
      parameters:
        - { $ref: '#/components/parameters/Name' }
        - { $ref: '#/components/parameters/Equipment' }
        - { $ref: '#/components/parameters/MinCapacity' }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
      responses:
        '200':
          description: 实验室列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LabPageResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /labs/{labId}:
    get:
      tags: [Labs]
      operationId: getLabDetail
      summary: 查询实验室详情
      description: Redis 详情缓存不可用时服务端回源 MySQL。
      parameters:
        - { $ref: '#/components/parameters/LabId' }
      responses:
        '200':
          description: 实验室详情
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LabResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /labs/{labId}/availability:
    get:
      tags: [Labs]
      operationId: getLabAvailability
      summary: 查询指定日期的实验室可用时段
      parameters:
        - { $ref: '#/components/parameters/LabId' }
        - name: date
          in: query
          required: true
          schema: { type: string, format: date }
        - name: from
          in: query
          schema: { type: string, pattern: '^([01]\d|2[0-3]):[0-5]\d$' }
        - name: to
          in: query
          schema: { type: string, pattern: '^([01]\d|2[0-3]):[0-5]\d$' }
      responses:
        '200':
          description: 开放时间及每个整点时隙的可用状态
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AvailabilityResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /admin/labs/{labId}:
    patch:
      tags: [Administration, Labs]
      operationId: updateLab
      summary: 管理员更新实验室白名单字段
      parameters:
        - { $ref: '#/components/parameters/LabId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LabUpdateRequest' }
      responses:
        '200':
          description: 更新后的实验室
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LabResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /chat/sessions:
    post:
      tags: [Chat]
      operationId: createChatSession
      summary: 为当前用户创建聊天会话
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChatSessionCreateRequest' }
      responses:
        '201':
          description: 已创建会话
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChatSessionResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /chat/sessions/{sessionId}/messages:
    post:
      tags: [Chat]
      operationId: sendChatMessage
      summary: 发送一条消息并获取一轮 Agent 结果
      parameters:
        - { $ref: '#/components/parameters/SessionId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChatMessageRequest' }
      responses:
        '200':
          description: Agent 回答，可包含待确认草案
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChatMessageResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableContent' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '504': { $ref: '#/components/responses/GatewayTimeout' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    LabId:
      name: labId
      in: path
      required: true
      schema: { type: string, maxLength: 32, example: LAB-B402 }
...
    SessionId:
      name: sessionId
      in: path
      required: true
      schema: { type: string, maxLength: 64 }
    Name:
      name: name
      in: query
      schema: { type: string, maxLength: 128 }
    Equipment:
      name: equipment
      in: query
      schema: { type: string, maxLength: 255 }
    MinCapacity:
      name: minCapacity
      in: query
      schema: { type: integer, minimum: 1 }
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
  responses:
    BadRequest: { description: 请求格式或字段校验失败, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    Unauthorized: { description: 未认证或登录凭据错误, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    Forbidden: { description: 无角色权限或资源不属于当前用户, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    NotFound: { description: 资源不存在或对当前用户不可见, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
...
    UnprocessableContent: { description: 可解析但不满足领域规则, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
...
    TooManyRequests: { description: 聊天请求超出每分钟限制, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    ServiceUnavailable: { description: Redis、模型或向量库不可用且不能安全降级, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
    GatewayTimeout: { description: 模型或向量库调用超时, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  schemas:
    SuccessEnvelope:
      type: object
      required: [code, message, data, requestId]
      properties:
        code: { type: integer, const: 200 }
        message: { type: string, const: 成功 }
        data: {}
        requestId: { type: string, example: req_01JABC }
    ErrorResponse:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          type: integer
          enum: [40000, 40001, 40100, 40300, 40301, 40400, 40401, 40402, 40900, 40901, 40902, 40903, 40904, 41300, 41500, 42200, 42201, 42202, 42203, 42204, 42205, 42900, 50000, 50300, 50301, 50400]
          example: 40901
        message: { type: string, example: 该实验室在指定时段已被预约，请重新选择。 }
        requestId: { type: string, example: req_01JABC }
...
    LabStatus: { type: string, enum: [ACTIVE, MAINTENANCE, DISABLED] }
    Lab:
      type: object
      required: [id, name, capacity, equipmentDescription, openTime, closeTime, status]
      properties:
        id: { type: string, example: LAB-B402 }
        name: { type: string }
        capacity: { type: integer, minimum: 1 }
        equipmentDescription: { type: string }
        openTime: { type: string, pattern: '^([01]\d|2[0-3]):[0-5]\d$' }
        closeTime: { type: string, pattern: '^([01]\d|2[0-3]):[0-5]\d$' }
        status: { $ref: '#/components/schemas/LabStatus' }
    PageLab:
      type: object
      required: [items, page, size, total]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Lab' } }
        page: { type: integer }
        size: { type: integer }
        total: { type: integer, format: int64 }
    LabPageResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties: { data: { $ref: '#/components/schemas/PageLab' } }
    LabResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties: { data: { $ref: '#/components/schemas/Lab' } }
    AvailabilitySlot:
      type: object
      required: [startTime, endTime, available]
      properties:
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        available: { type: boolean }
    AvailabilityData:
      type: object
      required: [labId, date, openTime, closeTime, slots]
      properties:
        labId: { type: string }
        date: { type: string, format: date }
        openTime: { type: string }
        closeTime: { type: string }
        slots: { type: array, items: { $ref: '#/components/schemas/AvailabilitySlot' } }
    AvailabilityResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties: { data: { $ref: '#/components/schemas/AvailabilityData' } }
...
    ActionType: { type: string, enum: [CREATE_RESERVATION, CANCEL_RESERVATION, CREATE_REPAIR_TICKET] }
    ActionDraft:
      type: object
      required: [actionId, actionType, sessionId, expiresAt, payload, notices]
      properties:
        actionId: { type: string, format: uuid }
        actionType: { $ref: '#/components/schemas/ActionType' }
        sessionId: { type: string }
        expiresAt: { type: string, format: date-time }
        payload: { type: object, additionalProperties: true }
        notices: { type: array, items: { type: string } }
...
    LabUpdateRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name: { type: string, minLength: 1, maxLength: 128 }
        capacity: { type: integer, minimum: 1 }
        equipmentDescription: { type: string, minLength: 1, maxLength: 10000 }
        openTime: { type: string, pattern: '^([01]\d|2[0-3]):[0-5]\d$' }
        closeTime: { type: string, pattern: '^([01]\d|2[0-3]):[0-5]\d$' }
        status: { $ref: '#/components/schemas/LabStatus' }
...
    Citation:
      type: object
      required: [documentTitle, version, chunkId, excerpt]
      properties:
        documentTitle: { type: string }
        version: { type: string }
        chunkId: { type: string }
        excerpt: { type: string }
...
    ChatSessionCreateRequest:
      type: object
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 128 }
    ChatSession:
      type: object
      required: [sessionId, createdAt]
      properties:
        sessionId: { type: string }
        name: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
    ChatSessionResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties: { data: { $ref: '#/components/schemas/ChatSession' } }
    ChatMessageRequest:
      type: object
      additionalProperties: false
      required: [content]
      properties:
        content: { type: string, minLength: 1, maxLength: 10000 }
    ChatMessage:
      type: object
      required: [sessionId, messageId, answer, citations, toolCallCount]
      properties:
        sessionId: { type: string }
        messageId: { type: string }
        answer: { type: string }
        citations: { type: array, items: { $ref: '#/components/schemas/Citation' } }
        toolCallCount: { type: integer, minimum: 0, maximum: 8 }
        draft: { $ref: '#/components/schemas/ActionDraft' }
    ChatMessageResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties: { data: { $ref: '#/components/schemas/ChatMessage' } }
...